refactored app to use camelCase file names
