+++ client/src/App.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from 'src/App';
+import App from 'src/app';
 
 test('renders learn react link', () => {
   render(<App />);
