+++ client/src/app.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import { BrowserRouter } from 'react-router-dom';
 
-import Routes from 'src/Routes';
+import Routes from 'src/routes';
 
 const App = () => {
   return (
