+++ client/src/components/loading/index.js
