+++ client/src/pages/organization/index.js
@@ -1,0 +1,17 @@
+import {
+  Container,
+  Row,
+  Col,
+} from 'reactstrap';
+
+const Organization = () => {
+  return (
+    <Container fluid>
+      <Row className='m-2'>
+        <Col className='text-center'><h2>Organization</h2></Col>
+      </Row>
+    </Container>
+  );
+}
+
+export default Organization;
