+++ client/src/pages/user/auth/register/index.js
@@ -17,7 +17,7 @@ import './index.css';
 import { publicAxios } from 'src/api/axios';
 import { register } from 'src/state/token/actions';
 import { clearMessage } from 'src/state/message/actions';
-import Error from 'src/components/Error';
+import Error from 'src/components/error';
 
 const CHECK_EMAIL_URL = 'check_email/'
 const MINIMUM_PASSWORD_LENGTH = 7;
