+++ client/src/routes.js
@@ -6,6 +6,7 @@ import Dashboard from 'src/pages/dashboard';
 import Login from 'src/pages/user/auth/login';
 import Register from 'src/pages/user/auth/register';
 import UserProfile from 'src/pages/user/profile';
+import Organization from 'src/pages/organization';
 
 const Routes = () => (
   <Switch>
@@ -14,8 +15,9 @@ const Routes = () => (
     <Route exact path='profile/' Component={UserProfile} />
     <Route Component={Layout}>
       <Route path='/' Component={Dashboard} />
-      {/* TODO: make a 404 page here */}
+      <Route path='organization/' Component={Organization} />
     </Route>
+    {/* TODO: make a 404 page here */}
   </Switch>
 );
